Move static methods to silence private reference warning
There is a lone problem marker in the Bndtools Eclipse workspace that reads:

`Export bndtools.central,  has 1,  private references [bndtools.model.repo]`

This is because RepositoryUtils exposed RepositoryBundle and
RepositoryBundleVersion as arguments to a couple static methods. This patch
moves those static methods to a new class in a non-exported package to
squelch the warning.

Signed-of-by: Sean Bright <sean.bright@gmail.com>
